Add navigation for pages
