+++ App.js
@@ -1,183 +1,83 @@
-import { StatusBar } from "expo-status-bar";
-import { CameraView, useCameraPermissions } from "expo-camera/next";
 import { useState } from "react";
 import { Button, StyleSheet, Text, Modal, Pressable, View } from "react-native";
+import { NavigationContainer } from "@react-navigation/native";
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import { createStackNavigator } from "@react-navigation/stack";
+import MaterialCommunityIcons from "react-native-vector-icons/MaterialCommunityIcons";
+import Inventory from "./Inventory";
+import Camera from "./Camera";
 
-export default function App() {
-  const [openCamera, setOpenCamera] = useState(false);
-  const [permission, requestPermission] = useCameraPermissions();
-  const [isFetchingData, setIsFetching] = useState(false);
-  const [scannedData, setScannedData] = useState(null);
-  const [modalVisible, setModalVisible] = useState(false);
-
-  async function startScanning() {
-    await requestPermission();
-    setOpenCamera(true);
-  }
-
+function SettingsScreen() {
   return (
-    <View style={styles.container}>
-      {!openCamera && <Button title="Scan Barcode" onPress={startScanning} />}
-      {openCamera && permission && (
-        <CameraView
-          style={styles.camera}
-          facing="back"
-          barcodeScannerSettings={{
-            barcodeTypes: ["upc_a"],
-          }}
-          onBarcodeScanned={(result) => {
-            if (!isFetchingData) {
-              setIsFetching(true);
-              let api =
-                "https://api.edamam.com/api/food-database/v2/parser?app_id=c2561828&app_key=62f3b9e585d90d6b875af08a312ca2d6&upc=" +
-                result.data;
-              fetch(api, {
-                methods: "GET",
-                headers: {
-                  "Content-Type": "application/json",
-                },
-              })
-                .then((response) => response.json())
-                .then((data) => {
-                  if (data && data["hints"]) {
-                    let scannedItem = {};
-                    let food = data["hints"][0]["food"];
-                    let nutrients = food["nutrients"];
-                    scannedItem["name"] = food["label"] ?? "Unknown";
-                    scannedItem["calories"] = nutrients["ENERC_KCAL"] ?? null;
-                    scannedItem["sugar"] = nutrients["SUGAR"] ?? null;
-                    scannedItem["fat"] = nutrients["FAT"] ?? null;
-                    scannedItem["weight"] =
-                      data["hints"][0]["measures"][1]["weight"] ?? null;
-                    setScannedData(scannedItem);
-                  } else {
-                    setScannedData(null);
-                  }
-                  setModalVisible(true);
-                  setOpenCamera(false);
-                  setIsFetching(false);
-                })
-                .catch(() => {
-                  setScannedData(null);
-                  setModalVisible(true);
-                  setOpenCamera(false);
-                  setIsFetching(false);
-                });
-            }
-          }}
-        ></CameraView>
-      )}
-      <Modal
-        animationType="slide"
-        transparent={true}
-        visible={modalVisible}
-        onRequestClose={() => {
-          setModalVisible(!modalVisible);
-        }}
-      >
-        <View style={styles.centeredView}>
-          <View style={styles.modalView}>
-            {scannedData && (
-              <View>
-                <Text style={styles.modalText}>
-                  Name: "{scannedData["name"]}"
-                </Text>
-                <Text style={styles.modalText}>
-                  Calories: {scannedData["calories"]}
-                </Text>
-                <Text style={styles.modalText}>
-                  Sugar: {scannedData["sugar"]}
-                </Text>
-                <Text style={styles.modalText}>Fat: {scannedData["fat"]}</Text>
-                <Text style={styles.modalText}>
-                  Weight: {scannedData["weight"]}
-                </Text>
-                <Pressable
-                  style={styles.button}
-                  onPress={() => setModalVisible(!modalVisible)}
-                >
-                  <Text style={styles.textStyle}>Cancel</Text>
-                </Pressable>
-                <Pressable
-                  style={styles.button}
-                  onPress={() => setModalVisible(!modalVisible)}
-                >
-                  <Text style={styles.textStyle}>Add to Inventory</Text>
-                </Pressable>
-              </View>
-            )}
-            {!scannedData && (
-              <View>
-                <Text style={styles.modalText}>
-                  There was an error getting the data. Most likely, the barcode
-                  scanned is not in our database. Please manually enter your
-                  food item.
-                </Text>
-                <Pressable
-                  style={styles.button}
-                  onPress={() => setModalVisible(!modalVisible)}
-                >
-                  <Text style={styles.textStyle}>Cancel</Text>
-                </Pressable>
-                <Pressable
-                  style={styles.button}
-                  onPress={() => setModalVisible(!modalVisible)}
-                >
-                  <Text style={styles.textStyle}>Manually Enter Data</Text>
-                </Pressable>
-              </View>
-            )}
-          </View>
-        </View>
-      </Modal>
+    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+      <Text>Settings!</Text>
     </View>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-  },
-  camera: {
-    flex: 1,
-  },
-  centeredView: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    marginTop: 22,
-  },
-  modalView: {
-    margin: 20,
-    backgroundColor: "white",
-    borderRadius: 20,
-    padding: 35,
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-    elevation: 5,
-  },
-  button: {
-    borderRadius: 20,
-    padding: 10,
-    elevation: 2,
-    marginTop: 10,
-    marginBottom: 10,
-    backgroundColor: "#2196F3",
-  },
-  textStyle: {
-    color: "white",
-    fontWeight: "bold",
-    textAlign: "center",
-  },
-  modalText: {
-    marginBottom: 15,
-    textAlign: "center",
-  },
-});
+const Tab = createBottomTabNavigator();
+const Stack = createStackNavigator();
+
+function HomeTabs() {
+  return (
+    <Tab.Navigator>
+      <Tab.Screen
+        name="Inventory"
+        component={Inventory}
+        options={{
+          tabBarLabel: "Inventory",
+          tabBarIcon: ({ color, size }) => (
+            <MaterialCommunityIcons
+              name="shopping-outline"
+              color={color}
+              size={size}
+            />
+          ),
+        }}
+      />
+      <Tab.Screen
+        name="Meals"
+        component={SettingsScreen}
+        options={{
+          tabBarLabel: "Meals",
+          tabBarIcon: ({ color, size }) => (
+            <MaterialCommunityIcons
+              name="food-outline"
+              color={color}
+              size={size}
+            />
+          ),
+        }}
+      />
+      <Tab.Screen
+        name="Settings"
+        component={SettingsScreen}
+        options={{
+          tabBarLabel: "Settings",
+          tabBarIcon: ({ color, size }) => (
+            <MaterialCommunityIcons
+              name="account-settings-outline"
+              color={color}
+              size={size}
+            />
+          ),
+        }}
+      />
+    </Tab.Navigator>
+  );
+}
+
+export default function App() {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen
+          name="Home"
+          component={HomeTabs}
+          options={{ headerShown: false }}
+        />
+        <Stack.Screen name="Barcode Scanner" component={Camera} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
