+++ package.json
@@ -9,13 +9,18 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^6.5.20",
+    "@react-navigation/stack": "^6.3.29",
     "expo": "~50.0.14",
     "expo-barcode-scanner": "~12.9.3",
     "expo-camera": "~14.1.1",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
     "react-native": "0.73.6",
-    "react-native-vision-camera": "^3.9.2"
+    "react-native-screens": "~3.29.0",
+    "react-native-vector-icons": "^10.0.3",
+    "react-native-vision-camera": "^3.9.2",
+    "react-native-gesture-handler": "~2.14.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
